Feat: Move deformable material to general isaaclab deformable asset
