+++ source/isaaclab/isaaclab/sim/__init__.pyi
@@ -53,6 +53,7 @@ __all__ = [
     "TriangleMeshSimplificationPropertiesCfg",
     "SpawnerCfg",
     "RigidObjectSpawnerCfg",
+    "DeformableObjectSpawnerCfg",
     "spawn_from_mjcf",
     "spawn_from_urdf",
     "spawn_from_usd",
@@ -71,8 +72,11 @@ __all__ = [
     "LightCfg",
     "SphereLightCfg",
     "spawn_rigid_body_material",
+    "spawn_deformable_body_material",
     "PhysicsMaterialCfg",
     "RigidBodyMaterialCfg",
+    "DeformableBodyMaterialCfg",
+    "SurfaceDeformableBodyMaterialCfg",
     "spawn_from_mdl_file",
     "spawn_preview_surface",
     "GlassMdlCfg",
@@ -218,6 +222,7 @@ from .schemas import (
 from .spawners import (
     SpawnerCfg,
     RigidObjectSpawnerCfg,
+    DeformableObjectSpawnerCfg,
     spawn_from_mjcf,
     spawn_from_urdf,
     spawn_from_usd,
@@ -236,8 +241,11 @@ from .spawners import (
     LightCfg,
     SphereLightCfg,
     spawn_rigid_body_material,
+    spawn_deformable_body_material,
     PhysicsMaterialCfg,
     RigidBodyMaterialCfg,
+    DeformableBodyMaterialCfg,
+    SurfaceDeformableBodyMaterialCfg,
     spawn_from_mdl_file,
     spawn_preview_surface,
     GlassMdlCfg,
