+++ source/isaaclab/isaaclab/sim/spawners/__init__.pyi
@@ -6,6 +6,7 @@
 __all__ = [
     "SpawnerCfg",
     "RigidObjectSpawnerCfg",
+    "DeformableObjectSpawnerCfg",
     "spawn_from_mjcf",
     "spawn_from_urdf",
     "spawn_from_usd",
@@ -24,8 +25,11 @@ __all__ = [
     "LightCfg",
     "SphereLightCfg",
     "spawn_rigid_body_material",
+    "spawn_deformable_body_material",
     "PhysicsMaterialCfg",
     "RigidBodyMaterialCfg",
+    "DeformableBodyMaterialCfg",
+    "SurfaceDeformableBodyMaterialCfg",
     "spawn_from_mdl_file",
     "spawn_preview_surface",
     "GlassMdlCfg",
@@ -65,7 +69,7 @@ __all__ = [
     "MultiUsdFileCfg",
 ]
 
-from .spawner_cfg import SpawnerCfg, RigidObjectSpawnerCfg
+from .spawner_cfg import SpawnerCfg, RigidObjectSpawnerCfg, DeformableObjectSpawnerCfg
 from .from_files import (
     spawn_from_mjcf,
     spawn_from_urdf,
@@ -89,6 +93,7 @@ from .lights import (
 )
 from .materials import (
     spawn_rigid_body_material,
+    spawn_deformable_body_material,
     PhysicsMaterialCfg,
     RigidBodyMaterialCfg,
     spawn_from_mdl_file,
@@ -97,6 +102,8 @@ from .materials import (
     MdlFileCfg,
     PreviewSurfaceCfg,
     VisualMaterialCfg,
+    DeformableBodyMaterialCfg,
+    SurfaceDeformableBodyMaterialCfg,
 )
 from .meshes import (
     spawn_mesh_capsule,
