+++ source/isaaclab/isaaclab/sim/spawners/materials/__init__.pyi
@@ -5,8 +5,11 @@
 
 __all__ = [
     "spawn_rigid_body_material",
+    "spawn_deformable_body_material",
     "PhysicsMaterialCfg",
     "RigidBodyMaterialCfg",
+    "DeformableBodyMaterialCfg",
+    "SurfaceDeformableBodyMaterialCfg",
     "spawn_from_mdl_file",
     "spawn_preview_surface",
     "GlassMdlCfg",
@@ -15,10 +18,12 @@ __all__ = [
     "VisualMaterialCfg",
 ]
 
-from .physics_materials import spawn_rigid_body_material
+from .physics_materials import spawn_rigid_body_material, spawn_deformable_body_material
 from .physics_materials_cfg import (
     PhysicsMaterialCfg,
     RigidBodyMaterialCfg,
+    DeformableBodyMaterialCfg,
+    SurfaceDeformableBodyMaterialCfg,
 )
 from .visual_materials import spawn_from_mdl_file, spawn_preview_surface
 from .visual_materials_cfg import GlassMdlCfg, MdlFileCfg, PreviewSurfaceCfg, VisualMaterialCfg
