+++ source/isaaclab/isaaclab/sim/spawners/spawner_cfg.py
@@ -100,3 +100,25 @@ class RigidObjectSpawnerCfg(SpawnerCfg):
 
     This adds the PhysxContactReporter API to all the rigid bodies in the given prim path and its children.
     """
+
+
+@configclass
+class DeformableObjectSpawnerCfg(SpawnerCfg):
+    """Configuration parameters for spawning a deformable asset.
+
+    Unlike rigid objects, deformable objects are affected by forces and can deform when subjected to
+    external forces. This class is used to configure the properties of the deformable object.
+
+    Deformable bodies don't have a separate collision mesh. The collision mesh is the same as the visual mesh.
+    The collision properties such as rest and collision offsets are specified in the :attr:`deformable_props`.
+
+    Note:
+        By default, all properties are set to None. This means that no properties will be added or modified
+        to the prim outside of the properties available by default when spawning the prim.
+    """
+
+    mass_props: schemas.MassPropertiesCfg | None = None
+    """Mass properties."""
+
+    deformable_props: schemas.DeformableBodyPropertiesCfg | None = None
+    """Deformable body properties. Only supported on PhysX backend for now."""
