--- source/isaaclab_physx/isaaclab_physx/sim/__init__.pyi
@@ -1,26 +1,0 @@
-# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-__all__ = [
-    "define_deformable_body_properties",
-    "modify_deformable_body_properties",
-    "DeformableBodyPropertiesCfg",
-    "DeformableObjectSpawnerCfg",
-    "spawn_deformable_body_material",
-    "DeformableBodyMaterialCfg",
-    "SurfaceDeformableBodyMaterialCfg",
-]
-
-from .schemas import (
-    define_deformable_body_properties,
-    modify_deformable_body_properties,
-    DeformableBodyPropertiesCfg
-)
-from .spawners import (
-    DeformableObjectSpawnerCfg,
-    spawn_deformable_body_material,
-    DeformableBodyMaterialCfg,
-    SurfaceDeformableBodyMaterialCfg,
-)
