--- source/isaaclab_physx/isaaclab_physx/sim/spawners/__init__.py
@@ -1,10 +1,0 @@
-# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-"""Sub-module containing utilities for creating prims in Omniverse for the PhysX backend."""
-
-from isaaclab.utils.module import lazy_export
-
-lazy_export()
