--- source/isaaclab_physx/isaaclab_physx/sim/spawners/__init__.pyi
@@ -1,18 +1,0 @@
-# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-__all__ = [
-    "DeformableObjectSpawnerCfg",
-    "spawn_deformable_body_material",
-    "DeformableBodyMaterialCfg",
-    "SurfaceDeformableBodyMaterialCfg",
-]
-
-from .spawner_cfg import DeformableObjectSpawnerCfg
-from .materials import (
-    spawn_deformable_body_material,
-    DeformableBodyMaterialCfg,
-    SurfaceDeformableBodyMaterialCfg,
-)
