+++ source/isaaclab_physx/setup.py
@@ -52,9 +52,6 @@ setup(
         "isaaclab_physx.sensors.contact_sensor",
         "isaaclab_physx.sensors.frame_transformer",
         "isaaclab_physx.sensors.imu",
-        "isaaclab_physx.sim",
-        "isaaclab_physx.sim.spawners",
-        "isaaclab_physx.sim.spawners.materials",
         "isaaclab_physx.test",
         "isaaclab_physx.test.benchmark",
         "isaaclab_physx.test.mock_interfaces",
